+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Privatize `Voice`, `VoiceLanguageInfo`, and `VoicesResponse` unnamed constructors.
 - Correct value type mapping of `VoicesResponse.toMap()`'s `voices` key.
+- Deprecate `VoicesResponse.getVoice()` method in favor of `VoicesResponse[]`.
 
 ## 1.0.0